# docs/swagger.yaml
basePath: /api/v1
definitions:
  common.ChangePassword:
    properties:
      new_password:
        type: string
      old_password:
        type: string
    type: object
...
  common.ResultWithErrors:
    properties:
      error:
        type: string
      result:
        type: string
    type: object
  database.Admin:
    properties:
      id:
        type: integer
      login:
        type: string
      password:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  title: Ruin keepers
  version: "0.1"
paths:
  /admin/changePassword:
    post:
      parameters:
      - description: Данные для смены пароля
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/common.ChangePassword'
      - description: Токен авторизации
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.ResultWithErrors'
      tags:
      - admin
  /admin/delete:
    post:
      parameters:
      - description: Токен авторизации
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.ResultWithErrors'
      tags:
      - admin
...
  /admin/login:
    post:
      parameters:
      - description: Данные админа
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/database.Admin'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.ResultWithErrors'
      tags:
      - admin
  /admin/register:
    post:
      parameters:
      - description: Данные админа
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/database.Admin'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.ResultWithErrors'
      tags:
      - admin
...
schemes:
- http
swagger: "2.0"